Add check if genome file doesn't contain raw sequence
Skip a replicon if there is no sequence associated with it,
and skip loading a genome project if it doesn't contain any valid replicons.

Note: Before this fix MicrobeDB would try and retrieve ALL replicons in the
database because it makes the assumption that it should be looked up if there is no
defined sequence given to it.