Removed the sequence columns from the schema and corresponding code that loads/accesses these columns.
Now the Replicon and Gene objects seek sequences from the flat files, fna then genbank, otherwise die with a failure.  This was to decrease database space needed.

As well the Genoneproject object contains a descructor that cleans up once a Genomeproject object goes out of scope, with regard to circular links between child objects such as Replicon and Gene objects.